+++ src/main/java/ua/edu/ucu/apps/flowerstore_web/manager/FlowerStore.java
@@ -1,11 +1,16 @@
-package ua.edu.ucu.apps.manager;
+package ua.edu.ucu.apps.flowerstore_web.manager;
 
-import ua.edu.ucu.apps.models.FlowerBucket;
-import ua.edu.ucu.apps.models.FlowerPack;
-import ua.edu.ucu.apps.models.FlowerType;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import ua.edu.ucu.apps.flowerstore_web.models.FlowerBucket;
+import ua.edu.ucu.apps.flowerstore_web.models.FlowerPack;
+import ua.edu.ucu.apps.flowerstore_web.models.FlowerType;
 
 import java.util.ArrayList;
 
+@SpringBootApplication
 public class FlowerStore {
     private final ArrayList<FlowerBucket> flowerBuckets = new ArrayList<>();
 
@@ -45,5 +50,10 @@ public class FlowerStore {
 
     public final void sell(final FlowerBucket flowerBucket) {
         flowerBuckets.remove(flowerBucket);
+    }
+
+    @GetMapping("/all")
+    public final ArrayList<FlowerBucket> getFlowerBuckets() {
+        return flowerBuckets;
     }
 }
