+++ src/main/java/ua/edu/ucu/apps/flowerstore_web/models/FlowerBucket.java
@@ -1,4 +1,4 @@
-package ua.edu.ucu.apps.models;
+package ua.edu.ucu.apps.flowerstore_web.models;
 
 import java.util.ArrayList;
