+++ src/main/java/ua/edu/ucu/apps/flowerstore_web/models/FlowerColor.java
@@ -1,4 +1,4 @@
-package ua.edu.ucu.apps.models;
+package ua.edu.ucu.apps.flowerstore_web.models;
 
 public enum FlowerColor {
     RED("#FF0000"),
