+++ src/main/java/ua/edu/ucu/apps/flowerstore_web/models/FlowerType.java
@@ -1,4 +1,4 @@
-package ua.edu.ucu.apps.models;
+package ua.edu.ucu.apps.flowerstore_web.models;
 
 public enum FlowerType {
     ROSE, CHAMOMILE, TULIP
