+++ src/main/java/ua/edu/ucu/apps/flowerstore_web/models/Rose.java
@@ -1,4 +1,4 @@
-package ua.edu.ucu.apps.models;
+package ua.edu.ucu.apps.flowerstore_web.models;
 
 public class Rose extends Flower {
     public Rose() {
