+++ src/main/java/ua/edu/ucu/apps/flowerstore_web/models/Tulip.java
@@ -1,4 +1,4 @@
-package ua.edu.ucu.apps.models;
+package ua.edu.ucu.apps.flowerstore_web.models;
 
 public class Tulip extends Flower {
     public Tulip() {
