+++ src/main/java/ua/edu/ucu/apps/flowerstore_web/models/package-info.java
@@ -1,4 +1,4 @@
-package ua.edu.ucu.apps.models;
+package ua.edu.ucu.apps.flowerstore_web.models;
 
 /*
  * This package contains all the models used in the application.
