+++ src/test/java/FlowerStoreTest.java
@@ -1,8 +1,8 @@
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import ua.edu.ucu.apps.manager.FlowerStore;
-import ua.edu.ucu.apps.models.*;
+import ua.edu.ucu.apps.flowerstore_web.manager.FlowerStore;
+import ua.edu.ucu.apps.flowerstore_web.models.*;
 
 import java.util.ArrayList;
